moved util to common directory
