+++ backend/clear_database.py
@@ -65,7 +65,7 @@ def clear_database():
         print("\n" + "="*60)
         print("⚠️  WARNING: This will DELETE ALL DATA in the database!")
         print("="*60)
-        from utilities.secure_logging import secure_print_database_url
+        from common.utilities.secure_logging import secure_print_database_url
         secure_print_database_url(settings.database_url)
         print("This action cannot be undone!")
         print("="*60)
