+++ backend/middleware/role_auth.py
@@ -5,7 +5,7 @@ from fastapi import HTTPException, status, Depends
 from sqlalchemy.orm import Session
 from database.connection import get_db
 from database.models import User
-from utilities.auth import get_current_user
+from common.utilities.auth import get_current_user
 
 def require_role(required_role: str):
     """
