+++ backend/services/async_chat_optimizer.py
@@ -13,7 +13,7 @@ import openai
 from functools import wraps
 
 from database.connection import settings
-from utilities.debug_logging import debug_log
+from common.utilities.debug_logging import debug_log
 
 # Performance monitoring
 logger = logging.getLogger(__name__)
