+++ backend/services/db_cache_service.py
@@ -10,7 +10,7 @@ from typing import Any, Dict, Optional, Union, List, Callable
 from datetime import datetime, timedelta
 from functools import wraps
 from sqlalchemy.orm import Session
-from utilities.redis_manager import redis_manager, cache_manager
+from common.utilities.redis_manager import redis_manager, cache_manager
 
 # Logger for database cache operations
 logger = logging.getLogger(__name__)
